Some cleanups. Basically removing never-used code.
* libopenvas/plugutils.c (is_shell_command_present): Removed.
It is not used anywhere except in
openvas-plugins/plugins/nmap_wrapper/nmap_wrapper.c for
very ancient NASL level.

* libopenvas/plugutils.h: Removed proto accordingly.

* libopenvas/popen.c (nessus_popen4): Removed unused
code that is deaactivcated with "#if 0".
(append_argv, destroy_argv): marked as to be deleted
eventually.

* libopenvas/network.c (nessus_print_SSL_certificate,
nessus_print_peer_SSL_certificate): Removed. It is unused
code that was deactivcated with "#if 0".
(nsend): Removed unused code that was deactivated with
"#if 0".



git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@2363 423fd1db-d629-0410-8442-d21db03e70f4