Added initial support for saving rules of users that were remotely authenticated. Rules file will be saved in OPENVAS_STATE_DIR/users-remote/ldap/USERNAME/auth/rules .
* misc/openvas_auth.c (openvas_auth_store_user_rules): Improved doc.
Expect a directory where to save the rules file (instead of just the
username), to circumvent discrimination of locally and remotely
authenticated users.

* misc/openvas_auth.h (openvas_auth_store_user_rules): Adjusted header.

* misc/ldap_auth.c: Include openvas_auth.
(ldap_auth_query_rules): New function, queries and saves the
access-rules of a user.
(ldap_authenticate): Call ldap_auth_query_rules for users and admins
to query and save the access-rules.

* doc/example.auth.conf: Added new configuration parameters.


git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@7365 423fd1db-d629-0410-8442-d21db03e70f4