Add envvar OPENVAS_DISABLE_LIBSSH
2012-10-22  Werner Koch  <wk@gnupg.org>

	* nasl/nasl_init.c (init_nasl_library) [HAVE_LIBSSH]: Define nasl
	constant "_HAVE_LIBSSH" by default.  Allow to disable it using
	OPENVAS_DISABLE_LIBSSH.  Remove use of envvar OPENVAS_USE_LIBSSH.
--

If NASL has been build with libssh support _HAVE_LIBSSH is now defined
by default.  To disable the use of libssh the envvar
OPENVAS_DISABLE_LIBSSH may now be used.  Note that this is only a
feature test macro for use in NVTs - the actual functions are
available even if macro is false.




git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@14302 423fd1db-d629-0410-8442-d21db03e70f4