Fix for pkcs#8 ssh keys and gnutls < 2.11.
2012-11-23  Werner Koch  <wk@gnupg.org>

	* nasl/nasl_ssh.c: Include gcrypt.h.
	(add_tlv): Fixup negative integer values.
	(mpi_from_gnutls_datum, mpi_to_gnutls_datum): New.  Only used with
	gnutls < 2.11.
	(pkcs8_to_sshprivatekey): Swap P and Q is needed and recompute u.
	Only used with gnutls < 2.11.
--

Note that I only used made up tests to check the new code.  A nicer
solution would be to do without using CRT so that we don't need p,q,
and u.  Libgcrypt supports this, but there is no way to tell libssh
that there are no values for p,q, and u.




git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@14600 423fd1db-d629-0410-8442-d21db03e70f4