Make libksba a mandatory dependency of openvas-libraries/nasl. Move and improve dependency check.
* CMakeLists.txt: Remove check for libksba. It is only used in
"nasl/", so it should be checked there.

* INSTALL: Move libksba from the "Recommended" to the general
dependencies section.

* nasl/CMakeLists.txt: Move dependency check here from top-level
CMakeLists.txt. Add version check with version mentioned in INSTALL.
Consider it an error if libksba is not found in the required version.
Do not make useless calls to ksba-config if ksba was not found or not
at the required version.

git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@26502 423fd1db-d629-0410-8442-d21db03e70f4