Fix #40: Check the version of libgcrypt
Previously CMake only checked for the presence of libgcrypt, causing
build failures when the installed libgcrypt was too old and did not
provide the required symbols.

This commit adds a version check for libgcrypt which will fail if the
version is below 1.6. This version requirement is now documented in the
INSTALL file as well.