Backport Scanner PR #171.
Fix parameter pollution in certain nasl functions.

Some NASL functions use get_var_by_name() instead of
get_local_var_by_name(), which "climbs" in the context when searching
for a variable ie. if a parameter 'foo' is not provided, the function
would use the variable 'foo' if found.

Unify the behaviour by replacing all get_local_var_by_name()
with get_var_by_name() calls, and change the later to not climb
in the nasl context when searching for the variable.