Introduce generic vtref struct and use for xref.
This introduces a struct for VT references.
The nvti objects is extended to have a list of the vtref
objects.
Basically the new struct is handled similar to the prefs.

Also, the xref element of nvti is removed and the API
uses the new list of vtrefs to mimic the previous behaviour.
Actually this is transitional until all modules will use
the new generic references.