Convert tag timestamps to seconds since epoch
When adding a tag named "last_modification" or "creation_date"
then assume the value is a timestamp and concert it to seconds
since epoch.

This integer value is then expressed as a string and added
as tag instead of the previous timestamp string.

The essential achievment is to internally only store seconds
since epoch for the NVT meta data.
The reiving module that processes the data should not know
about the timestamp format(s) used in the scanner(s).
Via OSP only seconds since epoch should be used for timestamps.