Ignore redundant tag cvss_base.
When a NASL script sets the tag "cvss_base" it will not be added
to the tags for the respective NVT anymore. Thus it will also not
be stored anymore in the redis database.

This saves some space and, when tags are parsed, computing time.
Not a fundamental improvement though.

The OSP protocol already ignores this tag because the value
can be computed from the vector anyway.
So, this does not require other changes.