add max_scan_hosts and max_alive_hosts scan restrictions
If max_scan_hosts number of alive hosts are reached no more alive
hosts are put on the alive detection queue. Instead the finish
signal is put on the queue.

If max_alive_hosts number of alive hosts are identified no more
alive hosts are identified and the sending of new pings is stopped.

If max_scan_hosts is reached we continue scanning for alive hosts
until max_alive_hosts was reached or no more hosts to scan.