send all not vuln scanned hosts to ospd-openvas
Previously only dead hosts were sent to ospd-openvas for calculating the progress bar.
Now dead hosts plus the hosts which are alive but were not vuln scanned by openvas are
included in the list of dead hosts to be sent to ospd-openvas as well.
Alive hosts may not be vuln scanned if max_scan_hosts is reached.