Use new redis connection for putting finish signal
put_finish_signal_on_queue() now uses a new redis
connection to main_db every time it is called.
This makes this function independent of internal
status of the scan. If for example the db of the
scanner struct is NULL for which ever reason, it
would not be possible to push the finish signal
on the queue.
This function is ideally only called once in
program lifetime. It may be called more than once
if errors occur or the thread gets cancelled.