Send the correct num of dead hosts to ospd-openvas
When we reach the max_scan_hosts limit we do not
send the true number of dead hosts.
The number needed by ospd-opevas for calculating
the correct progress bar and not ending a scan in
the interrupted state is calculated by subtracting
the number max_scan_hosts from the number
of target hosts if max_scan_hosts was reached.