Alternate between methods of alive detection
Now we do not e.g. send all icmp pings first.
If ICMP and TCP-SYN is chosen we continuously
send a few icmp pings then a few TCP_SYN pings
and so on.
This way we can achieve a a smooth progress bar
for ospd-openvas without the problem of needing
to correct the progress in the wrong direction if
a later alive test methods detected more results
than a previous one.