Change which alive hosts are added to hash table
Only add unique alive hosts which are also present
in the target list.
We use short circuit evaluation to not add hosts
in the alive hosts hash table which are not also
in the list of target hosts.
This way we make sure to only have the alive hosts
which we are looking for in our hash table of
alive host.