Handle scan restriction for last batch too
It could happen that during the last batch the
limit is reached. If the max_scan_hosts limit
got reached we can not send the true number of
dead hosts to ospd-openvas. We need to include
the alive hosts which are above the limit as
dead hosts for the progress bar.

E.g. curr_alive=110 prev_alive=90 max_scan_hosts=100 batch=100.
Normally we would send 80 as dead in this batch (20 new
alive hosts) but because of the restriction we send 90 as
dead. The 10 hosts which are over the limit are considered
as dead.