Use correct type size for port elements
The ports which are used for TCP pings are stored
in a GArray. TCP ports are of size uint16_t.
Now the array elements are all of size uint16_t.
When ports are added to the array they are first
cast to the appropriate type. The port lists
was already validated, so there are no values which
are bigger than uint16_t in it.