Make arp ping faster
We can reuse the libnet context if the interface
for a given packet is the same as the privously
used one.
We do not need to resolve the address here because
it is done somewhere else and saved in the
gvm_host_t.
If we already have a src address we do not need to
find one.