Fix: Require libpaho for mqtt support
MQTT support is required for openvas-scanner. It would be possible to
just not build the mqtt support in gvm-libs but in that case it would
be required to check in openvas-scanner if gvm-libs has mqtt support.
Therefore it is easier to always build gvm-libs with mqtt support.

Closes #694

(cherry picked from commit 7104b434635762ea598bb08b901248ee0fb5a882)