Add: Boreas timeout option (#711)
* Add scanner preference test_alive_wait_timeout.
It default to 3, because boreas had hardcoded three rounds of 1 second.

* Move timeout definition to another file

* Add function to get the timeout from preferences. If not set or the value is wrong, use the defined one.

* Use the new function for the wait timeout, instead of the hardcoded value

* Add timeout option to the client functions

* Wraps up the run_cli function with the timeout option, for backward compatibility.

What:
Make optional the waiting for replies timeout. This doesn't change the original behavior if not set.
Setting this in the openvas.conf, change the default value system wide.

Jira: SC-680

Why:
Some deveices or networks are too slow, and replies are not get in time, and therefore the target considered dead.

How:
Set the scanner preference (in GSA in the scan config, or in the openvas.conf configuration file). Run a scan against dead hosts and see the logs, how long it takes to finished the alive scan.

(cherry picked from commit 688fc637a06d3fea7f0e1275d45c5678fb0a089e)