Change: Set LDAP version to 3 after initialization (#723)
* Set LDAP version to 3 after initialization

LDAP version is by default set to 3, but when StarTLS
is disabled then the initialization is done for the second time.
And after the initialization version is set to 2. This change
is to set the version back to 3.

Signed-off-by: Franciszek Klajn <fklajn@opera.com>

* Fix coding style

Signed-off-by: Franciszek Klajn <fklajn@opera.com>
(cherry picked from commit b727c7e36db36dd9982b3160024518afcca19b15)