Adapt docker images
Pushes the version tags to docker hub on a tag release

- greenbone/gvm-libs:${major}
- greenbone/gvm-libs:${major.minor}
- greenbone/gvm-libs:${major.minor.patch}

This will allow container users to choose the newest 22 without
having to adapt to each minor or patch version manually. If the
latest 22 is buggy, they can jump back to either a patch or a minor
version until there is a new 22 version that fixes it.

When the tag version is the highest major version and is not the first
major release with minor and patch version is 0 than it also creates a
latest and stable tag.

Just to give an overview, here are some examples:

Release 22.5.1 on greenbone/gvm-libs would result in the
following tags:

- greenbone/gvm-libs:22
- greenbone/gvm-libs:22.5
- greenbone/gvm-libs:22.5.1
- greenbone/gvm-libs:latest
- greenbone/gvm-libs:stable

Release 23.0.0 on greenbone/gvm-libs would result in the
following tags:

- greenbone/gvm-libs:23
- greenbone/gvm-libs:23.0
- greenbone/gvm-libs:23.0.0

Release 23.0.1 on greenbone/gvm-libs would result in the
following tags:

- greenbone/gvm-libs:23
- greenbone/gvm-libs:23.0
- greenbone/gvm-libs:23.0.1
- greenbone/gvm-libs:latest
- greenbone/gvm-libs:stable

Release 22.5.2 when there is already a 23.0.1 version available would
result in the following tags:

- greenbone/gvm-libs:22
- greenbone/gvm-libs:22.5
- greenbone/gvm-libs:22.5.2