Change: Don't perform alive tests if "consider alive" method is selected
When more than one alive test method is selected, boreas check for
each one until the host is alive.
This will be done also even if the "consider alive" methods is selected.

This patch improves boreas to not run other selected alive methods
if "consider alive" method is set and there are no scan restrictions.

In case of scan restrictions (max_scan_host preference), this will not apply
to not fill the limited slots with probably dead hosts.