Add: Added conversion routines for different CPE notation formats.
Added routines to read a v2.2-conformant URI of a CPE or a
formatted string binding of a CPE into some kind of WFN struct.
Added routines to convert the WFN struct into a formatted string
binding or a CPE v2.2-conformant URI. Added routines to convert a
formatted string binding into a v2.2-conformant URI and vice versa.