Change: Add all tests to all target
Always rebuild the tests when something has changed. This is required
for getting the testing working with CMake and C in VSCode. Not having
the tests in the all target required to rebuild the tests manually when
something has change before the tests can be executed in the VSCode UI
again.