Change: Fallback to "*" section for all settings in a log domain
Use all log settings from the start section (our default domain) not
only selected ones if the setting is not set in a domain. For example
with the following config
```
[*]
prepend=%t %s %
prepend_time_format=%Y-%m-%d %Hh%M.%S %Z

[foo]
file=-
```

for the "foo" log domain the `prepend_time_format` from the star section
should be used too if it is not set. Before only the `prepend` string
was used.