Always free the MQTTResponse
Freeing the response only on error is probably OK given the way we're
using the API, but we might as well always free the response to be safe.

The documentation gives very little clue about precisely when to use
MQTTResponse_free.

I see some mallocs of properties in the WAIT_FOR_CONNACK case in
MQTTClient_connectURIVersion, which might indicate that on success there
is memory to be freed, at least for MQTTClient_connect5.