Clear client in mqtt_client_destroy
I think this is what the original author had in mind, because
it's a bit strange to assign NULL to the variable `client` only
to return immediately after.

There are also places that check client before using the mqtt,
so clearing client adds a little safety, in case anyone adds
code that uses an mqtt after running destroy.