Change: remove block end comments
The scanning block is small enough that it's easy to see where it ends.

I think the host_info one is trying to say that it was the last time `reader` is being used,
but it's hard to understand. If that's really needed then it would be better to enclose
`reader` in a separate block or function.