Chore: set fix debian codename in /etc/apt/*.sources
To prevent accidental distribution upgrades, e.g. when using gvm-libs as
a base, we should set the codename instead of rolling release.

This minimizes the effort of keeping track of debian releases manually,
checking always for whatever is stable while having a safe guard for
underlying images or runtime environments.