Change: Use libcurl4 with OpenSSL instead of GnuTLS
Require OpenSSL for using libcurl with SSL because it seems the GnuTLS
version can't handle certificates passed as blobs. This change affects
the following `.so` files of gvm-libs

```
libgvm_agent_controller.so
libgvm_http_scanner.so
libgvm_http.so
libgvm_openvasd.so
libgvm_container_image_scanner.so
```

All these `.so` files are now incompatible with GPLv2 only binaries
because OpenSSL >= version 3 is published under the Apache 2 license.