Fix: Ignore tests themselves for coverage
Add a configuration file for `gcovr` to ensure that files ending with
`_tests.c` or in the `tests/` directory are excluded when computing the
unit test coverage.

As the `_tests` files are certain to be executed fully during unit
tests, they all achieve a "coverage" of or close enough to 100 %, which
makes the total test coverage of the functional code (i.e., excluding
tests) seem larger than it actually is.